Replace all instances of djatwood/formailer with Rubonnek/formailer
Having issues vendoring the modules and deploying in gcloud. Getting
unknown syscalls due to an apparent Google Cloud Functions Framework not
being vendored correctly by go.

Attempting this fix to see if I can deploy my modified formailer (with
the Hbs handler) formailer as a second generation Google Cloud Function.

If you are reading this commit, it means the deployment was successful.